further organization by putting option order creation behind methods
